requirements.txt
# FraudLens Requirements
# Author: Yobie Benjamin
# Date: 2025-08-26 18:34:00 PDT

# Core dependencies
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
scikit-learn==1.5.1
certifi==2024.8.30

# Async and concurrency
asyncio-mqtt==0.16.2
aiofiles==24.1.0
aiohttp==3.10.5

# Configuration
pyyaml==6.0.2
toml==0.10.2
python-dotenv==1.0.1

# Resource monitoring
psutil==6.0.0
py-cpuinfo==9.0.0

# Apple Silicon optimization (optional, install if available)
# mlx==0.16.0  # Uncomment for Apple Silicon

# Model formats
onnx==1.16.2
onnxruntime==1.19.2
safetensors==0.4.4

# Image processing
Pillow==10.4.0
opencv-python==4.10.0.84

# PDF processing
pypdf==4.3.1
pdfplumber==0.11.4

# Text processing
nltk==3.9.1
transformers==4.44.2
tokenizers==0.19.1
sentencepiece==0.2.0

# Audio/Video processing (optional)
# librosa==0.10.2  # For audio
# moviepy==1.0.3  # For video

# ML frameworks (optional, choose based on needs)
# torch==2.4.1  # PyTorch
# tensorflow==2.17.0  # TensorFlow
# jax==0.4.31  # JAX

# CoreML support (macOS only)
# coremltools==7.2  # Uncomment for CoreML support

# API and web
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.0

# Gmail integration
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
Faker==20.0.0

# Code quality
black==24.8.0
ruff==0.6.4
flake8==6.1.0
isort==5.13.2
mypy==1.11.2
bandit==1.7.5
safety==3.0.0
pre-commit==3.8.0

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.32
mkdocstrings==0.25.2

# Logging and monitoring
loguru==0.7.2
rich==13.8.0
tqdm==4.66.5

# Data validation
jsonschema==4.23.0
marshmallow==3.22.0

# Development tools
ipython==8.12.3
jupyter==1.0.0
notebook==7.2.2
memory-profiler==0.61.0